made it so these frustum matrices were doubles

+++ src/cmd/click_list.cpp
@@ -57,7 +57,7 @@ UnitCollection * ClickList::requestIterator (int minX,int minY, int maxX, int ma
     GFXGetFrustumVars (true,&l,&r,&b,&t,&n,&f);
     GFXFrustum (frustmat, drivel,l*(-2.*minX/g_game.x_resolution+1) /*  *g_game.MouseSensitivityX*/,r*(2.*maxX/g_game.x_resolution-1)/*  *g_game.MouseSensitivityX*/,t*(-2.*minY/g_game.y_resolution+1) /*  *g_game.MouseSensitivityY*/,b*(2.*maxY/g_game.y_resolution-1)/*  *g_game.MouseSensitivityY*/,n,f);
     _Universe->AccessCamera()->GetView (view);
-    float frustum [6][4];
+    double frustum [6][4];
     GFXCalculateFrustum(frustum,view,frustmat);
     Unit * un;
     while ((un=myParent->current())) {
